more consistent context and property values
In the former examples, schema:url and schema:image were declared or used
in such a way that their values were interpreted as literals.
While the examples were not focused on this particular point,
I think it is bad practice to use broken (or at least unsual) data as example.